refactor: Fold executioner into StateBox
