+++ lib/index.js
@@ -6,7 +6,6 @@
 //   https://aws.amazon.com/step-functions/
 //   https://aws.amazon.com/blogs/aws/new-aws-step-functions-build-distributed-applications-using-visual-workstateMachines/
 
-const executioner = require('./executioner')
 const StateMachines = require('./state-machines')
 const Resources = require('./resources')
 const Status = require('./Status')
@@ -20,10 +19,13 @@ class Statebox {
     this.ready_ = findDao(options)
       .then(dao => {
         info(options.messages, 'Statebox is ready')
+        this.dao = dao
+        this.callbackManager = new CallbackManager()
         this.options.dao = dao
-        this.options.executioner = executioner
         this.options.stateMachines = this.stateMachines
-        this.options.callbackManager = new CallbackManager()
+        this.options.callbackManager = this.callbackManager
+        this.options.startExecution = (input, stateMachineName, executionOptions) =>
+          this.startExecution(input, stateMachineName, executionOptions)
       })
   }
 
@@ -91,20 +93,41 @@ class Statebox {
     return this.stateMachines.findStates(options)
   }
 
-  startExecution (input, stateMachineName, executionOptions) {
-    return executioner(
+  async startExecution (input, stateMachineName, executionOptions) {
+    // References
+    //   http://docs.aws.amazon.com/step-functions/latest/apireference/API_StartExecution.html
+    //   http://docs.aws.amazon.com/step-functions/latest/apireference/API_DescribeExecution.html
+    const stateMachineToExecute = this.stateMachines.findStateMachineByName(stateMachineName)
+    if (!stateMachineToExecute) {
+      // No stateMachine!
+      throw new Error(`Unknown stateMachine with name '${stateMachineName}'`)
+    } // if ...
+
+    const currentResource = stateMachineToExecute.definition.States[stateMachineToExecute.startAt].Resource
+    const executionDescription = await this.dao.createNewExecution(
+      stateMachineToExecute.startAt,
+      currentResource,
       input,
       stateMachineName,
-      executionOptions,
-      this.options
+      executionOptions
     )
-  } // startExecution
+
+    stateMachineToExecute.processState(executionDescription.executionName)
+    if (hasDelayedResponse(executionOptions)) {
+      return this.callbackManager.addCallback(
+        executionOptions.sendResponse,
+        executionDescription.executionName
+      )
+    }
+
+    return executionDescription
+  } // executioner
 
   async stopExecution (cause, errorCode, executionName, executionOptions) {
-    const executionDescription = await this.options.dao.findExecutionByName(executionName)
+    const executionDescription = await this.dao.findExecutionByName(executionName)
 
     if (executionDescription && executionDescription.status === Status.RUNNING) {
-      return this.options.dao.stopExecution(
+      return this.dao.stopExecution(
         cause,
         errorCode,
         executionName,
@@ -116,11 +139,11 @@ class Statebox {
   } // _stopExecution
 
   describeExecution (executionName) {
-    return this.options.dao.findExecutionByName(executionName)
+    return this.dao.findExecutionByName(executionName)
   } // describeExecution
 
   async sendTaskSuccess (executionName, output) {
-    const executionDescription = await this.options.dao.findExecutionByName(executionName)
+    const executionDescription = await this.dao.findExecutionByName(executionName)
 
     if (executionDescription && executionDescription.status === Status.RUNNING) {
       const stateToRun = this.stateMachines.findState(
@@ -134,7 +157,7 @@ class Statebox {
   } // sendTaskSuccess
 
   async sendTaskFailure (executionName, errorInfo) {
-    const executionDescription = await this.options.dao.findExecutionByName(executionName)
+    const executionDescription = await this.dao.findExecutionByName(executionName)
 
     if (executionDescription && executionDescription.status === Status.RUNNING) {
       const stateToRun = this.stateMachines.findState(
@@ -156,7 +179,7 @@ class Statebox {
   } // sendTaskLastHeartbeat
 
   async sendHeartbeat_ (executionName, updateResult, isLast) {
-    const executionDescription = await this.options.dao.findExecutionByName(executionName)
+    const executionDescription = await this.dao.findExecutionByName(executionName)
 
     if (executionDescription && executionDescription.status === Status.RUNNING) {
       const stateToRun = this.stateMachines.findState(
@@ -170,7 +193,7 @@ class Statebox {
   }
 
   async sendTaskRevivification (executionName) {
-    const executionDescription = await this.options.dao.findExecutionByName(executionName)
+    const executionDescription = await this.dao.findExecutionByName(executionName)
 
     if (executionDescription && executionDescription.status === Status.FAILED) {
       const stateMachine = this.stateMachines.findStateMachineByName(executionDescription.stateMachineName)
@@ -184,7 +207,7 @@ class Statebox {
     let notFound = 0
 
     do {
-      const executionDescription = await this.options.dao.findExecutionByName(executionName)
+      const executionDescription = await this.dao.findExecutionByName(executionName)
 
       if (typeof executionDescription !== 'object') {
         ++notFound
@@ -212,5 +235,10 @@ function info (messages, msg) {
     console.log(msg)
   }
 } // info
+
+function hasDelayedResponse (executionOptions) {
+  return executionOptions.sendResponse &&
+    executionOptions.sendResponse !== 'IMMEDIATELY'
+} // hasCallback
 
 module.exports = Statebox
