+++ lib/state-machines/state-types/Parent-base-state.js
@@ -6,7 +6,7 @@ class ParentBaseState extends BaseStateType {
   constructor (stateName, stateMachine, stateDefinition, options) {
     super(stateName, stateMachine, stateDefinition, options)
     this.options = options
-    this.executioner = options.executioner
+    this.startExecution = options.startExecution
   } // constructor
 
   makeChildExecution (
@@ -17,7 +17,7 @@ class ParentBaseState extends BaseStateType {
   ) {
     const branchContext = cloneDeep(input)
 
-    return this.executioner(
+    return this.startExecution(
       branchContext,
       stateMachineName,
       {
